Minor internal fix - removed usage of DWORD type

+++ include/vk_mem_alloc.h
@@ -3075,7 +3075,7 @@ static inline uint32_t VmaCountBitsSet(uint32_t v)
 static inline uint8_t VmaBitScanLSB(uint64_t mask)
 {
 #ifdef _MSC_VER
-    DWORD pos;
+    unsigned long pos;
     if (_BitScanForward64(&pos, mask))
         return static_cast<uint8_t>(pos);
 #else
@@ -3092,7 +3092,7 @@ static inline uint8_t VmaBitScanLSB(uint64_t mask)
 static inline uint8_t VmaBitScanLSB(uint32_t mask)
 {
 #ifdef _MSC_VER
-    DWORD pos;
+    unsigned long pos;
     if (_BitScanForward(&pos, mask))
         return static_cast<uint8_t>(pos);
 #else
